Make the i18n service available everywhere.
To get all dates localized, we need all runtimes to have the i18n
service, and since runtimes are made by the modulestores, we need to
pass it to them.

Soon (fingers crossed), modulestores won't be involved in making
runtimes, and many of these changes will go away.
