+++ common/lib/xmodule/xmodule/modulestore/django.py
@@ -6,14 +6,16 @@ Passes settings.MODULESTORE as kwargs to MongoModuleStore
 
 from __future__ import absolute_import
 from importlib import import_module
-
 import re
 
 from django.conf import settings
 from django.core.cache import get_cache, InvalidCacheBackendError
 from django.dispatch import Signal
+import django.utils
+
 from xmodule.modulestore.loc_mapper_store import LocMapperStore
 from xmodule.util.django import get_current_request_hostname
+
 
 # We may not always have the request_cache module available
 try:
@@ -38,7 +40,7 @@ def load_function(path):
     return getattr(import_module(module_path), name)
 
 
-def create_modulestore_instance(engine, doc_store_config, options):
+def create_modulestore_instance(engine, doc_store_config, options, i18n_service=None):
     """
     This will return a new instance of a modulestore given an engine and options
     """
@@ -68,6 +70,7 @@ def create_modulestore_instance(engine, doc_store_config, options):
         xblock_mixins=getattr(settings, 'XBLOCK_MIXINS', ()),
         xblock_select=getattr(settings, 'XBLOCK_SELECT_FUNCTION', None),
         doc_store_config=doc_store_config,
+        i18n_service=i18n_service or ModuleI18nService(),
         **_options
     )
 
@@ -186,3 +189,33 @@ def editable_modulestore(name='default'):
 
     else:
         return None
+
+
+class ModuleI18nService(object):
+    """
+    Implement the XBlock runtime "i18n" service.
+
+    Mostly a pass-through to Django's translation module.
+    django.utils.translation implements the gettext.Translations interface (it
+    has ugettext, ungettext, etc), so we can use it directly as the runtime
+    i18n service.
+
+    """
+    def __getattr__(self, name):
+        return getattr(django.utils.translation, name)
+
+    def strftime(self, *args, **kwargs):
+        """
+        A locale-aware implementation of strftime.
+        """
+        # This is the wrong place to import this function.  I'm putting it here
+        # because the xmodule test suite can't import this module, because
+        # Django is not available in that suite.  This function isn't called in
+        # that suite, so this hides the import so the test won't fail.
+        #
+        # As I said, this is wrong.  But Cale says this code will soon be
+        # refactored to a place that will be right, and the code can be made
+        # right there.  If you are reading this comment after April 1, 2014,
+        # then Cale was a liar.
+        from util.date_utils import strftime_localized
+        return strftime_localized(*args, **kwargs)
