+++ conf/locale/eo/LC_MESSAGES/djangojs.po
@@ -7,8 +7,8 @@ msgid ""
 msgstr ""
 "Project-Id-Version: 0.1a\n"
 "Report-Msgid-Bugs-To: openedx-translation@googlegroups.com\n"
-"POT-Creation-Date: 2014-02-19 15:38-0500\n"
-"PO-Revision-Date: 2014-02-19 20:39:50.876422\n"
+"POT-Creation-Date: 2014-02-20 13:01-0500\n"
+"PO-Revision-Date: 2014-02-20 18:02:05.375610\n"
 "Last-Translator: \n"
 "Language-Team: openedx-translation <openedx-translation@googlegroups.com>\n"
 "MIME-Version: 1.0\n"
