+++ lms/djangoapps/open_ended_grading/open_ended_notifications.py
@@ -7,6 +7,8 @@ from django.conf import settings
 from xmodule.open_ended_grading_classes import peer_grading_service
 from xmodule.open_ended_grading_classes.controller_query_service import ControllerQueryService
 
+from xmodule.modulestore.django import ModuleI18nService
+
 from courseware.access import has_access
 from lms.lib.xblock.runtime import LmsModuleSystem
 from edxmako.shortcuts import render_to_string
@@ -70,6 +72,9 @@ def peer_grading_notifications(course, user):
         render_template=render_to_string,
         replace_urls=None,
         descriptor_runtime=None,
+        services={
+            'i18n': ModuleI18nService(),
+        },
     )
     peer_gs = peer_grading_service.PeerGradingService(settings.OPEN_ENDED_GRADING_INTERFACE, system)
     pending_grading = False
@@ -131,6 +136,9 @@ def combined_notifications(course, user):
         render_template=render_to_string,
         replace_urls=None,
         descriptor_runtime=None,
+        services={
+            'i18n': ModuleI18nService(),
+        },
     )
     #Initialize controller query service using our mock system
     controller_qs = ControllerQueryService(settings.OPEN_ENDED_GRADING_INTERFACE, system)
